Put some classes in their own files.
