+++ icepaposc/dialog_settings.py
@@ -1,3 +1,22 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+# -----------------------------------------------------------------------------
+# This file is part of IcepapOCS link:
+#        https://github.com/ALBA-Synchrotron/IcepapOCS
+#
+# Copyright 2017:
+#       MAX IV Laboratory, Lund, Sweden
+#       CELLS / ALBA Synchrotron, Bellaterra, Spain
+#
+# Distributed under the terms of the GNU General Public License,
+# either version 3 of the License, or (at your option) any later version.
+# See LICENSE.txt for more info.
+#
+# You should have received a copy of the GNU General Public License
+# along with IcepapOCS. If not, see <http://www.gnu.org/licenses/>.
+# -----------------------------------------------------------------------------
+
 from PyQt4.QtGui import QDialog
 from PyQt4.QtGui import QDialogButtonBox
 from ui.ui_dialog_settings import Ui_DialogSettings
